many changes needs test, works on VM debug
Lost of changes to do with acceleration/ deceleration. Seems to work o.k. with 30 seconds ramp up time and ~ 15 secs ramp down. Extensive Visual Micro debug showed a call to the distancechecker routine took 30ms. The code is now incorporated into the main loop and takes < 1ms. All of this was done because the stepper stopped when anything# was sent over serial. The serial monitor on arduino may be the cause. Changing the serial baud rate did not seem to help, but setting it to a higher rate than 9600 would seem to be sensible anyway. Tests on the development machine do not show a motor stall when checking the number of millisecs from begin to end of void loop. Needs check on the observatory machine.