deceleration distance halved
When in field test on 15-5-19, POTH requests to slew over ran and this caused the stepper to oscillate around the target. So in the 'within 5 degrees ' routine, the stepper moveot distance has been halved in both directions from 2000 to 1000, so deceleration happens over half the distance it did previously. Needs test.