SIGNIFICANT CHANGES
many changes in response to bugs. Added a timer so that az is fetched from the encoder once every two seconds. The functions had not been constructed correctly so this has been remedied and tested  more testing required.
Also added a version number so this can be checked on the MCU lcd.