change to core to restructure and re use a routine
The code which forms the modulo of the difference between target and current Azimuth has been formed into and Integer routine. The new routine is called from two places - the LCDupdate routine where it is used to show distance to go on the LCD panel and also from WhichDirection which decide the shortest route to target. Needs test.