filename change to ino for editing in Arduino ide
I did this because the 4809 based changes for the new observatory are huge and will take a lot of time and testing. So the 4809 based stepper has its own branch.
Note that there may be empirical code changes in the observatory machine - eg acceleration values. The working versions of these changes are documented in a google docs file.