IMPORTANT update
It looks like the moveto target for the stepper was being achieved before the encoder had time to catch up. This was always a risk the way the stepper motor code has been written. It relies on stepper.moveto being a sufficiently large value that it does not get hit before the encoder angle matches the target azimuth (as the dome moves). So the moveto value has been increased to 150,000,000 which should be sufficient to allow the dome to reach the target azimuth. More testing required.