bugfix plus
The bug is if a Slewto azimuth 360 is requested, this causes lock up so trapped that and set a request for 360 to a request for 0 degrees. Also added a routine to check that the azimuth received via the SA command is valid.
There is a problem in the Pulsar dome drive (i.e. this routine) where the dome stops but over runs the target by about 5 degrees. So the decelvalue has been changed from 800 to 400, needs test.